Add linalg.lu_factor
This PR exposes `torch.lu` as `torch.linalg.lu_factor` and
`torch.linalg.lu_factor_ex`.

This PR also adds support for matrices with zero elements both in
the size of the matrix and the batch. Note that this function simply
returns empty tensors of the correct size in this case.

We add a test and an OpInfo for the new function.

This PR also adds documentation for this new function in line of
the documentation in the rest of `torch.linalg`.

ghstack-source-id: 2bf9fd22542e44dcb9725899e656802eaac96b48
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66933